+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support non-required link caching.
+- Support conditional links that do not have conditions and are not error handlers.
+
 ## [2.0.2] - 2026-02-15
 
 ### Fixed