+++ CHANGELOG.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Buffer inputs from optional links that arrive before the first execution.
 
+### Fixed
+
+- Protect input cache against parallel execution of the same `InputMergeActor` instance.
+
 ## [2.0.2] - 2026-02-15
 
 ### Fixed