+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [3.0.0rc1] - 2026-04-28
+
 ### Added
 
 - Support links that are explicitly optional (`required=False`).
@@ -140,7 +142,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Convert Ewoks `Graph` to graph of actors for exection.
 
-[unreleased]: https://github.com/ewoks-kit/ewoksppf/compare/v2.0.2...HEAD
+[unreleased]: https://github.com/ewoks-kit/ewoksppf/compare/v3.0.0rc1...HEAD
+[3.0.0rc1]: https://github.com/ewoks-kit/ewoksppf/compare/v2.0.2...v3.0.0rc1
 [2.0.2]: https://github.com/ewoks-kit/ewoksppf/compare/v2.0.1...v2.0.2
 [2.0.1]: https://github.com/ewoks-kit/ewoksppf/compare/v2.0.0...v2.0.1
 [2.0.0]: https://github.com/ewoks-kit/ewoksppf/compare/v1.3.0...v2.0.0